Add a sign command
This commit adds a new nabit command to sign an existing bag. The
thinking is that the bagit signing functionality could be useful for
people who are creating bags using different tools, and workflows.

I did spend some time looking into extracting the signature making
functionality into a separate utility. But the logic is complex enough
at this stage for it to continue to live with nabit for now. It also
seems to be well tested here, and there is not much sense in duplicating
that.